# Reuse BrinDesc and BrinRevmap in brininsert