This is a backport of upstream commit c1ec02be1d7 with some notable
adjustments:

(1) We don't backport the aminsertcleanup() AM routine, as that will
break ABI. Left merge fixmes with instructions for when that becomes
available.

~(2) Since we don't have REL14 commit c5b7ba4e67a, ExecOpenIndices() can
be called, even if ExecInsert() is never called (e.g. on QD or IIS with
no insert set). This means that ii_AmCache can be NULL, which means we
have to early return from index_insert_cleanup() if that is the case.
Left merge fixme with instructions for when that commit is absorbed.~

~(3) initialize_brin_insertstate() takes an isAO argument, like many
other BRIN routines. This avoids having to open the base relation to
figure out if this index is on an AO table or not.~

Note: the benefits of this patch is even more for GPDB, due to the added
AO specific fields in the structs involved.

Original commit message follows:

The brininsert code used to initialize (and destroy) BrinDesc and
BrinRevmap for each tuple, which is not free. This patch initializes
these structures only once, and reuses them for all inserts in the same
command. The data is passed through indexInfo->ii_AmCache.

This also introduces an optional AM callback "aminsertcleanup" that
allows performing custom cleanup in case simply pfree-ing ii_AmCache is
not sufficient (which is the case when the cache contains TupleDesc,
Buffers, and so on).

Author: Soumyadeep Chakraborty
Reviewed-by: Alvaro Herrera, Matthias van de Meent, Tomas Vondra
Discussion: https://postgr.es/m/CAE-ML%2B9r2%3DaO1wwji1sBN9gvPz2xRAtFUGfnffpd0ZqyuzjamA%40mail.gmail.com

Dev-pipeline: https://dev.ci.gpdb.pivotal.io/teams/main/pipelines/backport_brininsert?group=all