@soumyadeep2007 2023-11-26
Simple perf experiment with un-optimized build:

Setup: 
```
CREATE UNLOGGED TABLE heap (i int);
CREATE INDEX ON heap USING brin(i) WITH (pages_per_range=1);

CREATE UNLOGGED TABLE ao (i int) USING ao_row;
CREATE INDEX ON ao USING brin(i) WITH (pages_per_range=1);
```

Without patch:
```
INSERT INTO heap SELECT 1 FROM generate_series(1, 20000000);
Time: 39484.549 ms (00:39.485)

INSERT INTO ao SELECT 1 FROM generate_series(1, 20000000);
Time: 31829.162 ms (00:31.829)
```

With patch:
```
INSERT INTO heap SELECT 1 FROM generate_series(1, 20000000);
Time: 26939.779 ms (00:26.940)

INSERT INTO ao SELECT 1 FROM generate_series(1, 20000000);
Time: 23823.255 ms (00:23.823)
```