@haolinw 2023-11-27
> (3) initialize_brin_insertstate() takes an isAO argument, like many
other BRIN routines. This avoids having to open the base relation to
figure out if this index is on an AO table or not.

```c
static BrinInsertState *
initialize_brin_insertstate(Relation idxRel, IndexInfo *indexInfo, bool isAO)
{
	BrinInsertState *bistate;
	MemoryContext oldcxt;

	oldcxt = MemoryContextSwitchTo(indexInfo->ii_Context);
	bistate = palloc0(sizeof(BrinInsertState));
	bistate->bis_desc = brin_build_desc(idxRel);
	bistate->bis_rmAccess = brinRevmapInitialize(idxRel,
												 &bistate->bis_pages_per_range,
												 NULL);
	indexInfo->ii_AmCache = bistate;
	MemoryContextSwitchTo(oldcxt);

	return bistate;
}
```
I don't see `isAO` is used in current patch, is that intended for future extension of this function ?