@soumyadeep2007 2023-11-27
> I don't see isAO is used in current function, is that intended for future extension of this function ?

Ah good catch, I forgot to remove that. Removed, now and commit message adjusted.