@soumyadeep2007 2023-11-27
> (2) Since we don't have REL14 commit https://github.com/greenplum-db/gpdb/commit/c5b7ba4e67aeb5d6f824b74f94114d99ed6e42b7, ExecOpenIndices() can
be called, even if ExecInsert() is never called (e.g. on QD or IIS with
no insert set). This means that ii_AmCache can be NULL, which means we
have to early return from index_insert_cleanup() if that is the case.
Left merge fixme with instructions for when that commit is absorbed.

This aspect was fixed in upstream commit a82ee7ef3aacc2073d3ef5f4b7e5067fa08ea76c. So, I removed this workaround from the code and applied this commit instead.